libsnmp: Fix the overflow check in netsnmp_binary_array_get_subset()
Fixes: ccfcecbaf97c ("libsnmp: Fix a potential integer overflow in netsnmp_binary_array_get_subset()")

+++ snmplib/container_binary_array.c
@@ -641,7 +641,7 @@ netsnmp_binary_array_get_subset(netsnmp_container *c, void *key, int *len)
     }
 
     *len = end - start + 1;
-    if (*len <= 0 || *len > SIZE_MAX / sizeof(void*))
+    if (*len <= 0 || *len > INT_MAX / sizeof(void*))
         return NULL;
 
     subset = (void **)malloc((*len) * sizeof(void*));
